fix(ci): make pytest cross-shell, add tenacity, and fix unterminated string
- Replace multiline pytest command with single-line for Windows PowerShell compatibility
- Add tenacity>=8.2,<9 to requirements.txt for perspective module imports
- Fix f-string syntax errors in baselines.py by computing keys before formatting
- Remove problematic backslash line continuations in f-strings
- All changes maintain identical output behavior while fixing collection errors

+++ tests/test_baselines.py
@@ -290,8 +290,8 @@ class MockTokenizer:
     def __call__(self, text, **kwargs):
         return {
             "input_ids": torch.randint(1, 1000, (1, kwargs.get("max_length", 128))),
-            "attention_mask": torch.ones(1, kwargs.get("max_length", 128)),
-            "token_type_ids": torch.zeros(1, kwargs.get("max_length", 128)),
+            "attention_mask": torch.ones(1, kwargs.get("max_length", 128), dtype=torch.long),
+            "token_type_ids": torch.zeros(1, kwargs.get("max_length", 128), dtype=torch.long),
         }
 
     def save_pretrained(self, path):
